Split service to service interface and implementation
