+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/controller/CartItemController.java
@@ -3,6 +3,7 @@ package com.avdo.spring.app.controller;
 import com.avdo.spring.app.controller.dto.CreateCartItemDto;
 import com.avdo.spring.app.entity.CartItemEntity;
 import com.avdo.spring.app.service.CartItemService;
+import com.avdo.spring.app.service.CartItemServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
