+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/domain/model/CartItem.java
@@ -3,8 +3,10 @@ package com.avdo.spring.app.service.domain.model;
 import com.avdo.spring.app.entity.Cart;
 import com.avdo.spring.app.entity.Product;
 import lombok.Builder;
+import lombok.Data;
 
 @Builder
+@Data
 public class CartItem {
 
     private Long id;
